Move health tests.
