+++ contracts/health/tests/all_tests.rs
@@ -1,0 +1,1 @@
+mod tests;
