+++ contracts/health/tests/tests/helpers/mock_env.rs
@@ -22,7 +22,7 @@ use mars_types::{
     },
 };
 
-use crate::helpers::MockEnvBuilder;
+use super::MockEnvBuilder;
 
 pub struct MockEnv {
     pub app: BasicApp,
