+++ contracts/health/tests/tests/mod.rs
@@ -1,0 +1,8 @@
+mod helpers;
+
+mod test_health_state;
+mod test_health_values;
+mod test_hls;
+mod test_instantiate;
+mod test_liquidation_pricing;
+mod test_update_config;
